requirements.txt
python-dotenv==1.0.1
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.17
pinecone-client==5.0.0
langchain-pinecone==0.1.3
fastapi==0.11
uvicorn==0.30.1
streamlit==1.36.0
pypdf2==3.0.1